Refactor layer management code and use checkboxes
This commit refactors the layers into a single object at the top of
the code which contains the display name, the layer spec for mapbox,
and their default visibility. Each layer is then added to the map and
its default visibility is applied.

The list of links are replaced with checkboxes, which reflect the
default visibility of the layer and used as the source of the
visibility for the layer when clicked, rather than querying the map for
the layer visibility then inverting it (I think it might reduce lag a
bit).

I've also moved the event handler for the checkbox change event into a
separate function to avoid confusion around variable names and scope.