new Data Map and Dataset.
The  module now includes a new global map (`DATA_MAP`) which keeps the mapping of availble datasets to corresponding `Dataset` instances.
`Dataset` instances are dataclasses which load the data from (corresponding) data files in a **lazy** fashion. Data format supported so far are CSV and GeoJSON,
relying on pandas and geopandas, respectively. Therefore, `Dataset` instances include a `data` property actually mapping to data in the form of `DataFrame` or `GeoDataFrame`.