StandardiseDataset expanded with methods
dataset standardising functions are now methods of the Dataclass
encapsulating all the logic to standardise data.

At the moment, the StandardiseDataset includes a generated
property (i.e. std_data_ as in sklearn convention) to store
the standardised version of data.

In addition, the MERGE of two instances of StandardiseData has been
implemented.