Create master method to map requirements and create json
This file now implements three dataclasses which recursively
call eachother in order to create the overall json object for
the frontend.

Each Variable instance defines the requirements for that
variables to be transformed, and in Variables the data is
transformed.
One call should now create the entire json object for the front
end.