Add parsing of astersiks as NA in reading and na_false for large files
One file is known to have asterisks instead of NA's so I've added a
parameter to read_csv so that it is interpreted as an NA.
I have also added na_false for the large excel file as we know there are
no NA numbers and this speeds up reading slightly.

Some data is also commented out as it is not currently used on the map,
but we might want it very soon.