Bug fixes, read live data to list, and coerce data types to float
Added all the relevant instances of Datasets from standardised_datasets
to the initial lists.

Changed the list of live datasets so that the list is reading in the csvs
to dataframes rather than the paths which were then read in later. NB we will need to
find a better solution for this going forward, related to creation of the
dataset class.

Fixed the index column renaming so that columns are named consistently.

Coerced all columns to float64 due to issues with cols some being read as
strings.