SourceDataset: new property and (constructor) parameter for name. New stub for Dataset
The name of a SourceDataset will be automatically induced by the filename, if not provided.
The property (public API) returns the name of the SourceDataset as the _name attribute and
the corresponding resolution. This is to account for multiple dataset with same name
but different resolution.

Tests have been updated - all the instances of SourceDataset created by factory functions
have a unique name already setup (hard coded).

Initial stub implementation for (Standard)Dataset.