Reorganised into datasets with a data folder and removed cleaned files
To move towards the newer structure planned to be introduced later I
have reorganised all the files into a data folder within a datasets
folder. This allows for better encapsulation and less cross over of
scripts and data files.

In the process this has also deleted the cleaned data folder for static
data as the new process for loading this data in the MasterDataset class
in datasets.py does not require these files. As such they are redundant.