Reorganise run scripts for the phw covid scraper
This commit moves the script that was in the run_scrapers.py file
to a function in the package folder, and then creates a second file
to house a function that runs the scraping functions.

Note, the code content has changed very little, but it has been moved
and relative imports changed. Some string concatenation has also been
substituted with os.path.join.