Constants for LA and LSOA Count + Ref as Class variables + None Dataset removed
LA and LSOA counts are now constants in the module, so they can be easily changed.
Moreover, the `Dataset` class now includes two new Class Variables which are used to cache the (geo) dataframes of LA and LSOA data.
This should presumably reduce the time to load data from files.

Moreover, if a data source in Master Dataset is empty, it will be filtered out.
This is still to be handled properly when Variables are loaded (see #111)